Fixed a major issue with rotators
In some cases, you'd potentially hit two rotators with the same movement, for example:

```
- n
 @
- c

- c
 @
- N
```

When the N moves left, it would previously check the top `@` first (since it's loaded first), but it needs to check the bottom one first.

Solved by sorting bonds from the 'primary' element and then sorting modifiers from distances to that.