Properly rewrite the gravity system
Instead of checking for snakes that are supported by something, each iteration build up a list of everything supported (directly or indirectly) by a wall. This will properly handle mutually supported snakes without the hack I had earlier.

Example:

```
-----
-fed-
-10c-
-2ab-
-----
```

`0` is supported by `a` and `a` is supported by `0` at `e`.