Fix the distinction between cuts and spawns
Spawns are now the points where an atom starts and cuts are holes in the map. The difference is that the track can't go on either, but the final molecule is allowed to end up on a spawn.

This was enough to solve a few previous levels: Delta 6, Zeta 3, and Theta 3.