Implement validity checks + ordering
Validity checks: Use the previously mentioned simulate method to cut off impossible states early: If we've add too many toppings so that we can't possibly finish, that's invalid + if we block off a head so it can't reach any exit.
Ordering: Only expand the first head found (x/y loops) so we always finish in a specific order.